+++ ICD.Connect.Panels.CrestronPro/TriListAdapters/Dge/DmDge200CAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.Dge
+{
+	[KrangSettings("DmDge200C", typeof(DmDge200CAdapter))]
+	public sealed class DmDge200CAdapterSettings : AbstractDgeX00AdapterSettings
+	{
+	}
+}